Add style order-manager

+++ src/app/components/order-manager/order-manager.component.ts
@@ -41,7 +41,7 @@ export class OrderManagerComponent {
   selectedOrder = {
     _id: '',
     nroPedido: 0,
-    usuario: {usuario: '', password: '', email: '', rol: [''], cuil: '', nombre: '', fechaNacimiento: new Date(), direccion: '', telefono: ''},
+    usuario: { usuario: '', password: '', email: '', rol: [''], cuil: '', nombre: '', fechaNacimiento: new Date(), direccion: '', telefono: '' },
     productos: [],
     cantidad: [],
     subtotal: 0,
@@ -72,18 +72,22 @@ export class OrderManagerComponent {
     const orderToUpdate = this.pedidoService.pedidos.find(order => order.nroPedido === nroPedido);
     if (orderToUpdate) {
       this.pedidoService.updatePedido(orderToUpdate).subscribe(
-        () => {},
+        () => { },
         err => console.log(err)
       );
     }
   }
 
   deleteOrder(id: string) {
-    this.pedidoService.deletePedido(id).subscribe(
-      (res) => {
-        this.pedidoService.getAllPedido();
-      },
-      err => console.log(err)
-    )
+    const confirmDelete = window.confirm('¿Estás seguro de que deseas eliminar este pedido?');
+    if (confirmDelete) {
+      this.pedidoService.deletePedido(id).subscribe(
+        (res) => {
+          this.pedidoService.getAllPedido();
+        },
+        err => console.log(err)
+      )
+    }
   }
+
 }
